softdevice_controller: Optimize fetching of command complete
Store the command complete or status event inside a bt_buf
immediately. This saves RAM and a memcpy.

The tradeoff with this approach is that the input and output buffer
is shared. That is, they point to the same memory. Note that all
almost all commands that have both input and output parameters
have the same parameter as both input and output in the same memory
location. For a few commands this is not true:

- LE Encrypt. Here the SoftDevice Controller writes to the output
  parameter when the result is calculated.
- LE Read Local Resolvable Address. This command is not supported
- LE Set Extended Advertising Parameters. Selected TX Power is set after
  the advertising handle is read.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>